Relocate requisition and requisition line test to graphql crate
