+++ graphql/requisition/src/mutations/request_requisition/add_from_master_list.rs
@@ -123,3 +123,239 @@ impl MasterListNotFoundForThisStore {
         "Master list for this store is not found (might not be visible in this store)"
     }
 }
+
+#[cfg(test)]
+mod test {
+    use async_graphql::EmptyMutation;
+    use graphql_core::{
+        assert_graphql_query, assert_standard_graphql_error, test_helpers::setup_graphl_test,
+    };
+    use repository::{
+        mock::{
+            mock_request_draft_requisition, mock_sent_request_requisition_line, MockDataInserts,
+        },
+        RequisitionLine, StorageConnectionManager,
+    };
+    use serde_json::json;
+    use service::{
+        requisition::{
+            request_requisition::{
+                AddFromMasterList as ServiceInput, AddFromMasterListError as ServiceError,
+            },
+            RequisitionServiceTrait,
+        },
+        service_provider::{ServiceContext, ServiceProvider},
+    };
+
+    use crate::RequisitionMutations;
+
+    type DeleteLineMethod =
+        dyn Fn(&str, ServiceInput) -> Result<Vec<RequisitionLine>, ServiceError> + Sync + Send;
+
+    pub struct TestService(pub Box<DeleteLineMethod>);
+
+    impl RequisitionServiceTrait for TestService {
+        fn add_from_master_list(
+            &self,
+            _: &ServiceContext,
+            store_id: &str,
+            input: ServiceInput,
+        ) -> Result<Vec<RequisitionLine>, ServiceError> {
+            self.0(store_id, input)
+        }
+    }
+
+    fn service_provider(
+        test_service: TestService,
+        connection_manager: &StorageConnectionManager,
+    ) -> ServiceProvider {
+        let mut service_provider = ServiceProvider::new(connection_manager.clone());
+        service_provider.requisition_service = Box::new(test_service);
+        service_provider
+    }
+
+    fn empty_variables() -> serde_json::Value {
+        json!({
+          "input": {
+            "requestRequisitionId": "n/a",
+            "masterListId": "n/a",
+          },
+          "storeId": "n/a"
+        })
+    }
+
+    #[actix_rt::test]
+    async fn test_graphql_add_from_master_list_errors() {
+        let (_, _, connection_manager, settings) = setup_graphl_test(
+            EmptyMutation,
+            RequisitionMutations,
+            "test_graphql_add_from_master_list_structured_errors",
+            MockDataInserts::all(),
+        )
+        .await;
+
+        let mutation = r#"
+        mutation ($input: AddFromMasterListInput!, $storeId: String) {
+            addFromMasterList(storeId: $storeId, input: $input) {
+              ... on AddFromMasterListError {
+                error {
+                  __typename
+                }
+              }
+            }
+          }
+        "#;
+
+        // RequisitionDoesNotExist
+        let test_service = TestService(Box::new(|_, _| Err(ServiceError::RequisitionDoesNotExist)));
+
+        let expected = json!({
+            "addFromMasterList": {
+              "error": {
+                "__typename": "RecordDoesNotExist"
+              }
+            }
+          }
+        );
+
+        assert_graphql_query!(
+            &settings,
+            mutation,
+            &Some(empty_variables()),
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+
+        // CannotEditRequisition
+        let test_service = TestService(Box::new(|_, _| Err(ServiceError::CannotEditRequisition)));
+
+        let expected = json!({
+            "addFromMasterList": {
+              "error": {
+                "__typename": "CannotEditRequisition"
+              }
+            }
+          }
+        );
+
+        assert_graphql_query!(
+            &settings,
+            mutation,
+            &Some(empty_variables()),
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+
+        // MasterListNotFoundForThisStore
+        let test_service = TestService(Box::new(|_, _| {
+            Err(ServiceError::MasterListNotFoundForThisStore)
+        }));
+
+        let expected = json!({
+            "addFromMasterList": {
+              "error": {
+                "__typename": "MasterListNotFoundForThisStore"
+              }
+            }
+          }
+        );
+
+        assert_graphql_query!(
+            &settings,
+            mutation,
+            &Some(empty_variables()),
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+
+        // NotThisStoreRequisition
+        let test_service = TestService(Box::new(|_, _| Err(ServiceError::NotThisStoreRequisition)));
+        let expected_message = "Bad user input";
+        assert_standard_graphql_error!(
+            &settings,
+            &mutation,
+            &Some(empty_variables()),
+            &expected_message,
+            None,
+            Some(service_provider(test_service, &connection_manager))
+        );
+
+        // NotARequestRequisition
+        let test_service = TestService(Box::new(|_, _| Err(ServiceError::NotARequestRequisition)));
+        let expected_message = "Bad user input";
+        assert_standard_graphql_error!(
+            &settings,
+            &mutation,
+            &Some(empty_variables()),
+            &expected_message,
+            None,
+            Some(service_provider(test_service, &connection_manager))
+        );
+    }
+
+    #[actix_rt::test]
+    async fn test_graphql_add_from_master_list_success() {
+        let (_, _, connection_manager, settings) = setup_graphl_test(
+            EmptyMutation,
+            RequisitionMutations,
+            "test_graphql_add_from_master_list_success",
+            MockDataInserts::all(),
+        )
+        .await;
+
+        let mutation = r#"
+        mutation ($storeId: String, $input: AddFromMasterListInput!) {
+            addFromMasterList(storeId: $storeId, input: $input) {
+                ... on RequisitionLineConnector{
+                  nodes {
+                    id
+                  }
+                }
+            }
+          }
+        "#;
+
+        // Success
+        let test_service = TestService(Box::new(|store_id, input| {
+            assert_eq!(store_id, "store_a");
+            assert_eq!(
+                input,
+                ServiceInput {
+                    request_requisition_id: "id input".to_string(),
+                    master_list_id: "master list id input".to_string(),
+                }
+            );
+            Ok(vec![RequisitionLine {
+                requisition_line_row: mock_sent_request_requisition_line(),
+                requisition_row: mock_request_draft_requisition(),
+            }])
+        }));
+
+        let variables = json!({
+          "input": {
+            "requestRequisitionId": "id input",
+            "masterListId": "master list id input"
+          },
+          "storeId": "store_a"
+        });
+
+        let expected = json!({
+            "addFromMasterList": {
+              "nodes": [
+                {
+                  "id": mock_sent_request_requisition_line().id
+                }
+              ]
+            }
+          }
+        );
+
+        assert_graphql_query!(
+            &settings,
+            mutation,
+            &Some(variables),
+            &expected,
+            Some(service_provider(test_service, &connection_manager))
+        );
+    }
+}
