+++ graphql/requisition/src/query_tests/mod.rs
@@ -1,0 +1,4 @@
+mod requisition;
+mod requisition_by_number;
+mod requisition_loaders;
+mod requisitions;
