--- graphql/src/queries.rs
@@ -1,41 +1,0 @@
-use async_graphql::*;
-use graphql_core::pagination::PaginationInput;
-use graphql_requisition::requisition_queries::*;
-
-use graphql_types::types::*;
-
-#[derive(Default)]
-pub struct Queries;
-
-#[Object]
-impl Queries {
-    pub async fn requisition(
-        &self,
-        ctx: &Context<'_>,
-        store_id: String,
-        id: String,
-    ) -> Result<RequisitionResponse> {
-        get_requisition(ctx, &store_id, &id)
-    }
-
-    pub async fn requisitions(
-        &self,
-        ctx: &Context<'_>,
-        store_id: String,
-        page: Option<PaginationInput>,
-        filter: Option<RequisitionFilterInput>,
-        sort: Option<Vec<RequisitionSortInput>>,
-    ) -> Result<RequisitionsResponse> {
-        get_requisitions(ctx, &store_id, page, filter, sort)
-    }
-
-    pub async fn requisition_by_number(
-        &self,
-        ctx: &Context<'_>,
-        store_id: String,
-        requisition_number: u32,
-        r#type: RequisitionNodeType,
-    ) -> Result<RequisitionResponse> {
-        get_requisition_by_number(ctx, &store_id, requisition_number, r#type)
-    }
-}
