+++ server/src/test_utils.rs
@@ -1,8 +1,4 @@
-use repository::{
-    mock::{MockDataCollection, MockDataInserts},
-    test_db::{self, get_test_db_settings},
-    StorageConnection, StorageConnectionManager,
-};
+use repository::test_db::get_test_db_settings;
 
 use super::settings::{AuthSettings, ServerSettings, Settings, SyncSettings};
 
@@ -25,17 +21,4 @@ pub fn get_test_settings(db_name: &str) -> Settings {
             token_secret: "testtokensecret".to_string(),
         },
     }
-}
-
-pub async fn setup_all(
-    db_name: &str,
-    inserts: MockDataInserts,
-) -> (
-    MockDataCollection,
-    StorageConnection,
-    StorageConnectionManager,
-    Settings,
-) {
-    let repo = test_db::setup_all(db_name, inserts).await;
-    (repo.0, repo.1, repo.2, get_test_settings(db_name))
 }
