--- server/tests/graphql/requisition/mod.rs
@@ -1,8 +1,0 @@
-mod request_requisition;
-mod request_requisition_line;
-mod requisition;
-mod requisition_by_number;
-mod requisition_loaders;
-mod requisitions;
-mod response_requisition;
-mod response_requisition_line;
