--- server/tests/graphql/requisition/request_requisition/mod.rs
@@ -1,5 +1,0 @@
-mod add_from_master_list;
-mod delete;
-mod insert;
-mod update;
-mod use_suggested_quantity;
