--- server/tests/graphql/requisition/request_requisition/update.rs
@@ -1,220 +1,0 @@
-mod graphql {
-    use crate::graphql::{assert_graphql_query, assert_standard_graphql_error};
-    use repository::{
-        mock::{mock_name_a, mock_request_draft_requisition, MockDataInserts},
-        Requisition, StorageConnectionManager,
-    };
-    use serde_json::json;
-    use server::test_utils::setup_all;
-    use service::{
-        requisition::{
-            request_requisition::{
-                UpdateRequestRequisition as ServiceInput,
-                UpdateRequestRequisitionError as ServiceError, UpdateRequestRequstionStatus,
-            },
-            RequisitionServiceTrait,
-        },
-        service_provider::{ServiceContext, ServiceProvider},
-    };
-
-    type UpdateLineMethod =
-        dyn Fn(&str, ServiceInput) -> Result<Requisition, ServiceError> + Sync + Send;
-
-    pub struct TestService(pub Box<UpdateLineMethod>);
-
-    impl RequisitionServiceTrait for TestService {
-        fn update_request_requisition(
-            &self,
-            _: &ServiceContext,
-            store_id: &str,
-            input: ServiceInput,
-        ) -> Result<Requisition, ServiceError> {
-            self.0(store_id, input)
-        }
-    }
-
-    fn service_provider(
-        test_service: TestService,
-        connection_manager: &StorageConnectionManager,
-    ) -> ServiceProvider {
-        let mut service_provider = ServiceProvider::new(connection_manager.clone());
-        service_provider.requisition_service = Box::new(test_service);
-        service_provider
-    }
-
-    fn empty_variables() -> serde_json::Value {
-        json!({
-          "input": {
-            "id": "n/a",
-          },
-          "storeId": "n/a"
-        })
-    }
-
-    #[actix_rt::test]
-    async fn test_graphql_update_request_requisition_errors() {
-        let (_, _, connection_manager, settings) = setup_all(
-            "test_graphql_update_request_requisition_structured_errors",
-            MockDataInserts::all(),
-        )
-        .await;
-
-        let mutation = r#"
-        mutation ($input: UpdateRequestRequisitionInput!, $storeId: String) {
-            updateRequestRequisition(storeId: $storeId, input: $input) {
-              ... on UpdateRequestRequisitionError {
-                error {
-                  __typename
-                }
-              }
-            }
-          }
-        "#;
-
-        // RequisitionDoesNotExist
-        let test_service = TestService(Box::new(|_, _| Err(ServiceError::RequisitionDoesNotExist)));
-
-        let expected = json!({
-            "updateRequestRequisition": {
-              "error": {
-                "__typename": "RecordDoesNotExist"
-              }
-            }
-          }
-        );
-
-        assert_graphql_query!(
-            &settings,
-            mutation,
-            &Some(empty_variables()),
-            &expected,
-            Some(service_provider(test_service, &connection_manager))
-        );
-
-        // CannotEditRequisition
-        let test_service = TestService(Box::new(|_, _| Err(ServiceError::CannotEditRequisition)));
-
-        let expected = json!({
-            "updateRequestRequisition": {
-              "error": {
-                "__typename": "CannotEditRequisition"
-              }
-            }
-          }
-        );
-
-        assert_graphql_query!(
-            &settings,
-            mutation,
-            &Some(empty_variables()),
-            &expected,
-            Some(service_provider(test_service, &connection_manager))
-        );
-
-        // NotThisStoreRequisition
-        let test_service = TestService(Box::new(|_, _| Err(ServiceError::NotThisStoreRequisition)));
-        let expected_message = "Bad user input";
-        assert_standard_graphql_error!(
-            &settings,
-            &mutation,
-            &Some(empty_variables()),
-            &expected_message,
-            None,
-            Some(service_provider(test_service, &connection_manager))
-        );
-
-        // NotARequestRequisition
-        let test_service = TestService(Box::new(|_, _| Err(ServiceError::NotARequestRequisition)));
-        let expected_message = "Bad user input";
-        assert_standard_graphql_error!(
-            &settings,
-            &mutation,
-            &Some(empty_variables()),
-            &expected_message,
-            None,
-            Some(service_provider(test_service, &connection_manager))
-        );
-
-        // UpdatedRequisitionDoesNotExist
-        let test_service = TestService(Box::new(|_, _| {
-            Err(ServiceError::UpdatedRequisitionDoesNotExist)
-        }));
-        let expected_message = "Internal error";
-        assert_standard_graphql_error!(
-            &settings,
-            &mutation,
-            &Some(empty_variables()),
-            &expected_message,
-            None,
-            Some(service_provider(test_service, &connection_manager))
-        );
-    }
-
-    #[actix_rt::test]
-    async fn test_graphql_update_request_requisition_success() {
-        let (_, _, connection_manager, settings) = setup_all(
-            "test_graphql_update_request_requisition_success",
-            MockDataInserts::all(),
-        )
-        .await;
-
-        let mutation = r#"
-        mutation ($storeId: String, $input: UpdateRequestRequisitionInput!) {
-            updateRequestRequisition(storeId: $storeId, input: $input) {
-                ... on RequisitionNode {
-                    id
-                }
-            }
-          }
-        "#;
-
-        // Success
-        let test_service = TestService(Box::new(|store_id, input| {
-            assert_eq!(store_id, "store_a");
-            assert_eq!(
-                input,
-                ServiceInput {
-                    id: "id input".to_string(),
-                    colour: Some("colour input".to_string()),
-                    their_reference: Some("reference input".to_string()),
-                    comment: Some("comment input".to_string()),
-                    max_months_of_stock: Some(1.0),
-                    min_months_of_stock: Some(2.0),
-                    status: Some(UpdateRequestRequstionStatus::Sent)
-                }
-            );
-            Ok(Requisition {
-                requisition_row: mock_request_draft_requisition(),
-                name_row: mock_name_a(),
-            })
-        }));
-
-        let variables = json!({
-          "input": {
-            "id": "id input",
-            "maxMonthsOfStock": 1,
-            "minMonthsOfStock": 2,
-            "colour": "colour input",
-            "theirReference": "reference input",
-            "comment": "comment input",
-            "status": "SENT"
-          },
-          "storeId": "store_a"
-        });
-
-        let expected = json!({
-            "updateRequestRequisition": {
-                "id": mock_request_draft_requisition().id
-            }
-          }
-        );
-
-        assert_graphql_query!(
-            &settings,
-            mutation,
-            &Some(variables),
-            &expected,
-            Some(service_provider(test_service, &connection_manager))
-        );
-    }
-}
