--- server/tests/graphql/requisition/request_requisition/use_suggested_quantity.rs
@@ -1,201 +1,0 @@
-mod graphql {
-    use crate::graphql::{assert_graphql_query, assert_standard_graphql_error};
-    use repository::{
-        mock::{
-            mock_request_draft_requisition, mock_sent_request_requisition_line, MockDataInserts,
-        },
-        RequisitionLine, StorageConnectionManager,
-    };
-    use serde_json::json;
-    use server::test_utils::setup_all;
-    use service::{
-        requisition::{
-            request_requisition::{
-                UseSuggestedQuantity as ServiceInput, UseSuggestedQuantityError as ServiceError,
-            },
-            RequisitionServiceTrait,
-        },
-        service_provider::{ServiceContext, ServiceProvider},
-    };
-
-    type DeleteLineMethod =
-        dyn Fn(&str, ServiceInput) -> Result<Vec<RequisitionLine>, ServiceError> + Sync + Send;
-
-    pub struct TestService(pub Box<DeleteLineMethod>);
-
-    impl RequisitionServiceTrait for TestService {
-        fn use_suggested_quantity(
-            &self,
-            _: &ServiceContext,
-            store_id: &str,
-            input: ServiceInput,
-        ) -> Result<Vec<RequisitionLine>, ServiceError> {
-            self.0(store_id, input)
-        }
-    }
-
-    fn service_provider(
-        test_service: TestService,
-        connection_manager: &StorageConnectionManager,
-    ) -> ServiceProvider {
-        let mut service_provider = ServiceProvider::new(connection_manager.clone());
-        service_provider.requisition_service = Box::new(test_service);
-        service_provider
-    }
-
-    fn empty_variables() -> serde_json::Value {
-        json!({
-          "input": {
-            "requestRequisitionId": "n/a"
-          },
-          "storeId": "n/a"
-        })
-    }
-
-    #[actix_rt::test]
-    async fn test_graphql_use_suggested_quantity_errors() {
-        let (_, _, connection_manager, settings) = setup_all(
-            "test_graphql_use_suggested_quantity_structured_errors",
-            MockDataInserts::all(),
-        )
-        .await;
-
-        let mutation = r#"
-        mutation ($input: UseSuggestedQuantityInput!, $storeId: String) {
-            useSuggestedQuantity(storeId: $storeId, input: $input) {
-              ... on UseSuggestedQuantityError {
-                error {
-                  __typename
-                }
-              }
-            }
-          }
-        "#;
-
-        // RequisitionDoesNotExist
-        let test_service = TestService(Box::new(|_, _| Err(ServiceError::RequisitionDoesNotExist)));
-
-        let expected = json!({
-            "useSuggestedQuantity": {
-              "error": {
-                "__typename": "RecordDoesNotExist"
-              }
-            }
-          }
-        );
-
-        assert_graphql_query!(
-            &settings,
-            mutation,
-            &Some(empty_variables()),
-            &expected,
-            Some(service_provider(test_service, &connection_manager))
-        );
-
-        // CannotEditRequisition
-        let test_service = TestService(Box::new(|_, _| Err(ServiceError::CannotEditRequisition)));
-
-        let expected = json!({
-            "useSuggestedQuantity": {
-              "error": {
-                "__typename": "CannotEditRequisition"
-              }
-            }
-          }
-        );
-
-        assert_graphql_query!(
-            &settings,
-            mutation,
-            &Some(empty_variables()),
-            &expected,
-            Some(service_provider(test_service, &connection_manager))
-        );
-
-        // NotThisStoreRequisition
-        let test_service = TestService(Box::new(|_, _| Err(ServiceError::NotThisStoreRequisition)));
-        let expected_message = "Bad user input";
-        assert_standard_graphql_error!(
-            &settings,
-            &mutation,
-            &Some(empty_variables()),
-            &expected_message,
-            None,
-            Some(service_provider(test_service, &connection_manager))
-        );
-
-        // NotARequestRequisition
-        let test_service = TestService(Box::new(|_, _| Err(ServiceError::NotARequestRequisition)));
-        let expected_message = "Bad user input";
-        assert_standard_graphql_error!(
-            &settings,
-            &mutation,
-            &Some(empty_variables()),
-            &expected_message,
-            None,
-            Some(service_provider(test_service, &connection_manager))
-        );
-    }
-
-    #[actix_rt::test]
-    async fn test_graphql_use_suggested_quantity_success() {
-        let (_, _, connection_manager, settings) = setup_all(
-            "test_graphql_use_suggested_quantity_success",
-            MockDataInserts::all(),
-        )
-        .await;
-
-        let mutation = r#"
-        mutation ($storeId: String, $input: UseSuggestedQuantityInput!) {
-            useSuggestedQuantity(storeId: $storeId, input: $input) {
-                ... on RequisitionLineConnector{
-                  nodes {
-                    id
-                  }
-                }
-            }
-          }
-        "#;
-
-        // Success
-        let test_service = TestService(Box::new(|store_id, input| {
-            assert_eq!(store_id, "store_a");
-            assert_eq!(
-                input,
-                ServiceInput {
-                    request_requisition_id: "id input".to_string(),
-                }
-            );
-            Ok(vec![RequisitionLine {
-                requisition_line_row: mock_sent_request_requisition_line(),
-                requisition_row: mock_request_draft_requisition(),
-            }])
-        }));
-
-        let variables = json!({
-          "input": {
-            "requestRequisitionId": "id input"
-          },
-          "storeId": "store_a"
-        });
-
-        let expected = json!({
-            "useSuggestedQuantity": {
-              "nodes": [
-                {
-                  "id": mock_sent_request_requisition_line().id
-                }
-              ]
-            }
-          }
-        );
-
-        assert_graphql_query!(
-            &settings,
-            mutation,
-            &Some(variables),
-            &expected,
-            Some(service_provider(test_service, &connection_manager))
-        );
-    }
-}
