--- server/tests/graphql/requisition/request_requisition_line/mod.rs
@@ -1,3 +1,0 @@
-mod delete;
-mod insert;
-mod update;
