--- server/tests/graphql/requisition/response_requisition/mod.rs
@@ -1,3 +1,0 @@
-mod create_requisition_shipment;
-mod supply_requested_quantity;
-mod update;
